Usability fix for melee and ranged attacks.
When a character goes out of initiative or needs to do a coup de grace, radio buttons were used after the attack details -- making that confusing and error prone.  People forget to use the radio buttons when they need to.

Now, the radio buttons are relocated so that they appear first and people will be less likely to forget.