Hero Lab sync indicators for token properties now read-only
Previously due to sloppy programming on my part the "label" feature was used to indicate that attacks and equipment was imported from Hero Lab, providing the ability to synchronize if the data changed in MapTool or Hero Lab.

This "label" field is supposed to be used by users as a comment field to help them remember the purpose of those items, mainly attacks.

Now, the Hero Lab sync status is maintained in a read-only field and the "label" field can now be used by users as a comment field.