Removed obsolete token type
All PC and NPC tokens, even when imported from Hero Lab, are the "Basic" token type -- and this has been the case for at least two years.

Note:  This is potentially a breaking change.  Tokens using the old Hero Lab token type need to be assigned to the "Basic" token type.