Erase NPC Token Names
Created a macro to enable GMs to replace all NPC token names on a map with a name of the form "Unknown <Random Number>".  The new name is not guaranteed to be unique, so GMs will have to check their maps and fix any conflicts.  The GM name for the token is unchanged.