Fixed spells not showing for Hero Lab imports
Hero Lab data does not include spells per day information for spells in the XML even though it is in the HTML.  I had a "feature" that would hide spells if the per day was 0 for that level spell.  Due to this problem, spells are now always displayed UNLESS the per day for that level is less than 0 (as a way of hiding them, if you want, because you don't have them yet).

For Hero Lab users, they will be warned about the spells per day issue so they know how to fix it.  They will also now see all of the spells for editing and casting.