Fixed adjustments add/edit dialog for link items and special conditions
Fixed adjustments add/edit dialog for link items and special conditions.  Special conditions should not have been a checkbox.  Instead, a label (indicator) was needed to let you know if there was a special condition associated with the link item (for example an equipment item -- say it has a curse).

There is still a problem to be fixed.  The edit version of the dialog only reports whether link items had a special condition at the time the adjustment was added -- this needs to be fixed.