Added feature to allow treasure items to be stored in "treasure tokens".
You must use the Edit Token dialog on the "Config" tab to change the "Properties" to "Treasure".  Then, you can use the "Treasure (Token-based)" macros to add, edit, delete, and transfer to party treasure treasure items for a treasure token.  This enables you to store the treasure items at the map locations where they are found in tokens (and you can hide those tokens or not).

To make these treasure items visible to the party, you will have to edit them in the global treasure list using the "Treasure (Party and Global)" macros and change their "Treasure Status".