Add feature to allow HTML or text stat blocks to be cut-and-pasted for tokens
To enable GMs to do things more on-the-fly, I've added the ability to cut-and-paste a stat block for a token (NPC, PC, or whatever).  You don't have to enter any other data.  I recommend you assign an image to your token in the standard MapTool way.

If it turns out you need to track resources, such as HP, you can enter that information for the token as you go.

Note:  This also opens the door for much more easily using this framework to play other RPG games.