Added feature to flip all PC or Hero Lab tokens in the X or Y direction
This feature helps with doing in-person tabletop.  There's a bug here that will be fixed soon.  The intent is to flip all PC and NPC tokens, but by default all other map objects are treated as NPC tokens.  So, to deal with that temporarily, only PC and Hero Lab synchronized tokens will be flipped.